Add SAS256 and change enrollement handling
Add support for SAS256 and construct the right SAS in this case.

Enrollement handling now handles the case if the user already enrolled with
the PBX. In that case the ZRTP functions use a new information code# to inform
the application about the facts. The user may select "Don't trust anymore" and
delete the trust relationship with the PBX.