TIVI - Add Android support and add a Java interface
The modified clients/tivi/CMakeLists.txt configures makefiles for Andorid
support. Because Android uses Java I've added a Java interface to test the
CtZrtpSession class and it's callbacks. This was done with SWIG.

The CMakeLists.txt calls SWIG to generate the interface files. A small Java
class is available to test the interface. This Java class behaves like the
testdriver C++ process.