Fix a timing problem when handling SIP/SDES and ZRTP
A second, delayed SIP 200OK message could trigger a second SDES parse function
call. Usually this second call does nothing because it checks if we are
already in secure state. However, when starting ZRTP we set to another state
and the second function call goes thru. This could happen if we have a bad
network.

I added another check to prevent this. Once the SDES stream is in its active
state then dismiss any function call to create/parse/mix another SDES crypto
string for the existing SDES stream.

The SDES stream stays in active state while the call is active. The code uses
the two boolean variables 'useSdesForMedia' and 'useZrtpTunnel' to control
usage of the SDES stream.