Use new SIDH public key functions to ZRTP, add unit tests
This is a quite large commit which adds my SIDH fork as submodule and thus has all the
update/enhancements to build SIDH with cmake. The older SIDH code was removed from ZRTP
and the new SIDH functions were added.

Some new unit tests run and test this new code.