Fix handling when writing shared secrets to ZID cache
According to RFC6189, sections 4.6.1, 4.6.1.1 must update their RS1 retained shared secrets
in their caches. Only in case of a cache mismatch the RS1 update must be delayed until explicit
user verification (SAS verification), see 4.6.1.1.

However, due to a wrong initialization of a variable the protocol stack only saved the RS1 when
the user verified the SAS, thus usually only once. Due to the missing continuous update of RS1
the security/correctness of the key continuity was broken.

This bug was never part of the master branch, only in develop and some feature branches, thus never
officially released.

The fix changes this:
- the ZRTP stack now always writes a new RS1 to the ZID cache as required in 4.6.1
- if a cache mismatch happens, then the ZRTP stack does not write the RS1, only if user verifies the SAS again

Due to the misbehaviour the stack could report wrong cache mismatch, in particular if only one party
verified the SAS while the other party does not.