Improve tivi info string for Hello Hash
We provide to the tivi client a string description of the state of the
ZRTP Hello Hash.  On a ZRTP call this can be "Good", if it matched,
"Bad", if it didn't match, or "None", if it wasn't provided.  However
if the call isn't supposed to use ZRTP, i.e. it's using SDES and no
ZRTP Hello was received, then we provide a different info string.

We were providing the string "No hello".  This is a bit difficult to
interpret.  Let's say "No ZRTP" instead, as that better captures the
situation.