Implement emscripten support (by Chris Double)
This change was copied from ZrtpRandom.cpp and slightly adapted

Random number entropy is made via changes to ZRTPCPP to call
JavaScript Web Crypto routines. In the case of browser based clients
where libzina is run on a worker thread we can't call Web
Crypto. Clients may, for example, create an object containing
functions that obtain entropy from a static array. The worker thread
would manage the data in this array and when it gets low post a
message to the main thread asking for more data and refill it from the
reply. The main thread would use Web Crypto to get this data.

Unfortunately, `window.crypto` isn't in scope when run in a EM_ASM_INT
block from within a webbrowser. We work around this by accessing a
`Module.browserCrypto` object.  This object must be made available by
the in-browser JS code with something like:

    Module.browserCrypto = window.crypto;